Rewrite assembly dependency resolution logic as CustomPackageAction
